Add graphql-parser, implement GraphQLPrinter
Tried to add platform specific graphqlQueryToAst based on graphql-js and
graphql-java but while graphql-js generates an AST similar to that of
the graphql schema, graphql-java's parser creates a different representation.
Although this could be converted to the the same repesentation as provided by
graphql-js, instead I went with adding
https://github.com/kgbier/graphql-parser-kotlin and having a kotlin mpp
implementation of a graphqle query parsing.

I could not manage to make it build as a maven package, so for now I just
copied the sources from git. Also fixed parsing typeDeferred (created PR
for it in original project) and added hints in our implementation to aid
debugging the parsers.