Add processGraphql implementation
For now processGraphql replaces __everything selections with the simple fields
(scalars) of the given type recursively. It works by taking the graphql query
and the corresponding graphql schema.

Note: tried to use korio to load test files from the filesyste and while it
work in JVM the imports don't work in JS. Needs more investigation as
korio seems to be a really nice lib to use.